requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.9
Django==4.2.7
django-decouple==2.1
django-environ==0.11.2
django-js-asset==2.2.0
django-tinymce==4.1.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
